+++ packages/multichain-account-service/src/providers/SolAccountProvider.test.ts
@@ -116,7 +116,7 @@ function setup({
   const keyring = new MockSolanaKeyring(accounts);
 
   messenger.registerActionHandler(
-    'AccountsController:listMultichainAccounts',
+    'AccountsController:getAccounts',
     () => accounts,
   );
 
@@ -194,6 +194,22 @@ describe('SolAccountProvider', () => {
     expect(() => provider.getAccount(unknownAccount.id)).toThrow(
       `Unable to find account: ${unknownAccount.id}`,
     );
+  });
+
+  it('returns true if an account is compatible', () => {
+    const account = MOCK_SOL_ACCOUNT_1;
+    const { provider } = setup({
+      accounts: [account],
+    });
+    expect(provider.isAccountCompatible(account)).toBe(true);
+  });
+
+  it('returns false if an account is not compatible', () => {
+    const account = MOCK_HD_ACCOUNT_1;
+    const { provider } = setup({
+      accounts: [account],
+    });
+    expect(provider.isAccountCompatible(account)).toBe(false);
   });
 
   it('creates accounts', async () => {