+++ packages/multichain-account-service/src/tests/messenger.ts
@@ -47,6 +47,9 @@ export function getMultichainAccountServiceMessenger(
       'KeyringController:addNewKeyring',
       'NetworkController:findNetworkClientIdByChainId',
       'NetworkController:getNetworkClientById',
+      'KeyringController:createNewVaultAndKeychain',
+      'KeyringController:createNewVaultAndRestore',
+      'AccountsController:getAccounts',
     ],
   });
 }